Refactor : 3주차 리뷰 반영_#01
