+++ frontend/src/containers/IssueDetailPage/DetailBody/CommentElement/CommentElement.module.css
