+++ frontend/src/services/issue.js
@@ -55,13 +55,13 @@ export const patchIssueStatus = async (issueId, status) => {
   }
 };
 
-export const patchIssueTitle = async (issueId, issueTitle) => {
+export const patchIssueTitle = async (issueId, inputValue) => {
   try {
     const response = await customFetch({
       path: `/issue/${issueId}/title`,
       method: 'PATCH',
       body: {
-        title: issueTitle,
+        title: inputValue,
       },
     });
